+++ tests/test_06_math/test_626_local_bspline_interpolation.py
@@ -2,27 +2,33 @@
 # License: MIT License
 import pytest
 from ezdxf.math import Vector
-from ezdxf.math.bspline import(
-    local_cubic_bspline_interpolation, estimate_cubic_spline_tangents_5p, estimate_cubic_spline_tangents_3p,
-)
+from ezdxf.math.bspline import local_cubic_bspline_interpolation_from_tangents, local_cubic_bspline_interpolation
+from ezdxf.math import estimate_tangents
 
 POINTS1 = [(1, 1), (2, 4), (4, 1), (7, 6)]
 POINTS2 = [(1, 1), (2, 4), (4, 1), (7, 6), (5, 8), (3, 3), (1, 7)]
 
 
 def test_estimate_tangents_3p():
-    tangents = estimate_cubic_spline_tangents_3p(Vector.list(POINTS1))
+    tangents = estimate_tangents(Vector.list(POINTS1), method='3-points')
     assert len(tangents) == 4
 
 
 def test_estimate_tangents_5p():
-    tangents = estimate_cubic_spline_tangents_5p(Vector.list(POINTS1))
+    tangents = estimate_tangents(Vector.list(POINTS1), method='5-points')
     assert len(tangents) == 4
 
 
-def _test_local_cubic_bspline_interpolation():
+def test_local_cubic_bspline_interpolation_from_tangents():
     points = Vector.list(POINTS1)
-    tangents = estimate_cubic_spline_tangents_5p(points)
-    control_points, knots = local_cubic_bspline_interpolation(points, tangents)
-    assert len(control_points) == 4
-    assert len(knots) == 10
+    tangents = estimate_tangents(points)
+    control_points, knots = local_cubic_bspline_interpolation_from_tangents(points, tangents)
+    assert len(control_points) == 8
+    assert len(knots) == 8 + 4  # count + order
+
+
+def test_local_cubic_bspline_interpolation():
+    bspline = local_cubic_bspline_interpolation(POINTS1, method='5-points')
+    assert bspline.degree == 3
+    assert len(bspline.control_points) == 8
+    assert len(bspline.knots()) == 8 + 4  # count + order