+++ NEWS.md
@@ -47,6 +47,7 @@ Version 0.13b3 - dev
 - NEW: `ezdxf.math.cubic_bezier_from_ellipse()` returns an approximation for an elliptic arc by multiple cubic Bézier curves
 - NEW: `ezdxf.math.cubic_bezier_interpolation()` returns an interpolation curve for arbitrary data points as multiple cubic Bézier curves
 - NEW: `ezdxf.math.LUDecomposition` linear equation solver, for more linear algebra tools see module `ezdxf.math.linalg`
+- NEW: `ezdxf.math.local_cubic_bspline_interpolation()`
 - NEW: `ezdxf.render.random_2d_path()` generate random 2D path for testing purpose
 - NEW: `ezdxf.render.random_3d_path()` generate random 3D path for testing purpose
 - NEW: `BSpline.from_arc()` B-spline from an arc, best approximation with a minimum number of control points